TW-86069 Fix jackson dependency
The problem has been caused by our patched library mixed with a provided dependency of jackson. We've activated annotation-driven spring injection and that causes some jackson Modules to be loaded by Spring for all `BaseControllers`. If jsr310 is present, it will break, marking it as a Unnamed Module.
In order to prevent spring from running those again, we can ensure the plugin uses the same classloader.